Package upgrades (#61)

+++ tests/pkgchk-cli.tests/GithubTests.fs
@@ -31,7 +31,7 @@ module GithubTests =
     let issueClient () = Substitute.For<IIssuesClient>()
 
     let issueGet (issue: Issue) (issueClient: IIssuesClient) =
-        issueClient.Get(Arg.Any<string>(), Arg.Any<string>(), Arg.Any<int>()).Returns(issue)
+        issueClient.Get(Arg.Any<string>(), Arg.Any<string>(), Arg.Any<int64>()).Returns(issue)
         |> ignore
 
         issueClient
@@ -43,7 +43,7 @@ module GithubTests =
     let commentClient () = Substitute.For<IIssueCommentsClient>()
 
     let commentsGet (comments: IssueComment[]) (client: IIssueCommentsClient) =
-        client.GetAllForIssue(Arg.Any<string>(), Arg.Any<string>(), Arg.Any<int>()).Returns(comments)
+        client.GetAllForIssue(Arg.Any<string>(), Arg.Any<string>(), Arg.Any<int64>()).Returns(comments)
         |> ignore
 
         client
@@ -61,7 +61,7 @@ module GithubTests =
         let issueClient = issueClient () |> bindComments commentClient
         let client = client () |> bindIssues issueClient
 
-        issueClient.Get(Arg.Any<string>(), Arg.Any<string>(), Arg.Any<int>()).Returns(throwIssueException)
+        issueClient.Get(Arg.Any<string>(), Arg.Any<string>(), Arg.Any<int64>()).Returns(throwIssueException)
         |> ignore
 
         let rt = pkgchk.Github.getIssueComments client repo 1
